jeirizai-task8/src/main/java/com/epam/rd/java/basic/practice8/db/entity/Team.java
package com.epam.rd.java.basic.practice8.db.entity;

import java.util.Objects;

public class Team {

    private int id;
    private String name;

    public int getId() {
        return id;
    }

    public void setId(Integer id) {
        this.id = id;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public static Team createTeam(String name) {
        Team team = new Team();
        team.setName(name);
        return team;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) {
            return true;
        }
        if (o == null) {
            return false;
        }
        if (!(o instanceof Team)) {
            return false;
        }
        Team other = (Team) o;
        if (name == null) {
            if (other.name != null) {
                return false;
            }
        } else if (!name.equals(other.name)) {
            return false;
        }
        return true;
    }

    @Override
    public int hashCode() {
        return Objects.hash(getName());
    }

    @Override
    public String toString() {
        return name;
    }
}
